+++ morpheus/core/xml_functions.h
@@ -25,7 +25,7 @@
 
 
 /// Some methods making consistent handling for reading and writing values and XML structure
-
+string readFile(string filename);
 XMLNode parseXMLFile(string filename, XMLResults *pResults);
 XMLNode getXMLNode(const XMLNode XML_base, const string& path);
 string getXMLPath(const XMLNode node);